refactor: Remove deprecated AI constants and update imports
- Deleted the ai.ts constants file as it is no longer needed.
- Updated import paths in ai-service.ts, index.ts, base-post-generator.ts, and groq-provider.ts to reflect the new structure.
- Adjusted the GroqProvider class to utilize DEFAULT_CONFIG for model and token settings.
- Added TODO comments for Gemini and OpenAI providers indicating they are under development.
