+++ constants/ai/ai.ts
