+++ constants/ai/api/client.ts
@@ -1,0 +1,6 @@
+const getGenerationsConfig = {
+    page: 1,
+    limit: 10
+}
+
+export { getGenerationsConfig };
