+++ lib/ai/index.ts
@@ -2,7 +2,7 @@ import { AIService } from './ai-service';
 import type { AIProviderType } from './factory/provider-factory';
 import type { AIServiceConfig } from '@/types/ai';
 export * from '@/types/ai';
-export * from '@/constants/ai';
+export * from '@/constants/ai/ai';
 export { AIService };
 
 // Factory function for easy instantiation
